requirements.txt
python-telegram-bot==13.9
requests
pytest
allure-pytest
psycopg2